Split glass-islands into modules
